// src/chat/intents/intent.types.ts
export enum ChatIntent {
  INVENTORY_AVAILABILITY = 'INVENTORY_AVAILABILITY',
  INVENTORY_SEARCH = 'INVENTORY_SEARCH',
  PAYMENT_ESTIMATE = 'PAYMENT_ESTIMATE',
  PRICING_REQUEST = 'PRICING_REQUEST',
  SERVICE_APPOINTMENT = 'SERVICE_APPOINTMENT',
  HOURS_LOCATION = 'HOURS_LOCATION',
  HUMAN_HANDOFF = 'HUMAN_HANDOFF',
  GENERAL_QUESTION = 'GENERAL_QUESTION',
  HOLD_VEHICLE = 'HOLD_VEHICLE',
  INTELLIGENT_QUERY = 'INTELLIGENT_QUERY',
  PARTS_INQUIRY = 'PARTS_INQUIRY',
}